Make the markup even more valid
By default, pygments was wrapping its output in a div. We override it and use a
span, because cgit will put it in <code> which allows only phrasing content,
which span is but div isn't.

We tell the span to display as a block because otherwise it does weird things
to line height and causes lines to be desynced from line numbers(?).